Member Domain 추가
